+++ Source/Plugins/ParallelPortDevice/ppcapturethread.h
@@ -1,5 +1,5 @@
 //parallelportplugin
-//Copyright (C) 2013  Sven Gijsen
+//Copyright (C) 2014  Sven Gijsen
 //
 //This file is part of StimulGL.
 //StimulGL is free software: you can redistribute it and/or modify
@@ -50,7 +50,7 @@ private:
 	ParallelPort *portDev;
 
 signals:
-	void recieveThreadTriggered(short Value);
+	void recieveThreadTriggered(int Value);
 	void recieveThreadStarted(QString timeTextStamp);
 	void recieveThreadStopped(QString timeTextStamp);
 };
